refactor(ds): unify HTTP edge DS-recomputation on canonical binary_truth frame
The HTTP POST /api/v1/edges path had its own parallel CDST implementation
(`trigger_edge_ds_recomputation` + `recompute_claim_belief`) that wrote to
a separate `graph-topology` frame with hardcoded math
(`mass = source_truth * 0.7 * 0.5`) and handled only 4 of the 13 epistemic
relationships. After PR #170, the canonical edge-factor algorithm lives in
`epigraph-mcp::tools::ds_auto` and uses `binary_truth` + the
`RestrictionKind` framework.

This collapses the duplicate:

- Extract `auto_wire_ds_for_edge`, `auto_wire_edge_if_epistemic`,
  `EdgeFactorOutcome`, the `recompute_combined_belief` cascade, and the
  small helpers (`ensure_binary_frame`, `binary_frame`, `mass_to_json`,
  `parse_stored_bba`, `focal_to_key`) into `epigraph_engine::edge_factor`.
- Re-export from `epigraph-mcp::tools::ds_auto` so existing MCP call sites
  (PR #170/#171) keep working unchanged.
- Rewrite `routes/edges.rs::trigger_edge_ds_recomputation` to delegate to
  the engine module. Fetches source claim's `agent_id` for BBA attribution
  (matching legacy semantics). Keeps `propagate_to_dependents` in api as
  HTTP-only orchestration; its inner `recompute_claim_belief` now calls
  `epigraph_engine::edge_factor::recompute_claim_belief_binary`.
- Drop the `is_evidential_relationship` gate at the HTTP call site — the
  engine's `RestrictionKind::Neutral` self-filter covers the same set and
  adds 9 more epistemic relationships (refines/elaborates/specializes/
  generalizes/supersedes/informs/undercuts/rebuts/frame_validates).

Net: 78 insertions, 470 deletions. Two parallel algorithms become one.

The 62 `graph-topology` BBAs in prod (last write 2026-05-16) are orphaned
by this change; followup will null them or migrate.
