+++ crates/epigraph-engine/src/lib.rs
@@ -28,6 +28,7 @@ pub mod counterfactual;
 pub mod dag;
 pub mod diverse_select;
 pub mod domain_decay;
+pub mod edge_factor;
 pub mod embedder;
 pub mod epistemic_interval;
 pub mod error_mass;
