+++ crates/epigraph-mcp/src/tools/ds_auto.rs
@@ -14,11 +14,13 @@ use uuid::Uuid;
 
 use epigraph_db::{FrameRepository, MassFunctionRepository, PerspectiveRepository};
 use epigraph_ds::{combination, measures, FocalElement, FrameOfDiscernment, MassFunction};
-use epigraph_engine::epistemic_interval::{
-    restrict_epistemic_frame_evidence, restrict_epistemic_negative, restrict_epistemic_positive,
-    EpistemicInterval,
+
+// Edge-factor auto-wire moved to `epigraph_engine::edge_factor` so the HTTP
+// route layer can share a single algorithm. Re-export keeps the existing
+// MCP call sites (`tools::ingestion`, `tools::workflows`) working unchanged.
+pub use epigraph_engine::edge_factor::{
+    auto_wire_ds_for_edge, auto_wire_edge_if_epistemic, EdgeFactorOutcome,
 };
-use epigraph_engine::sheaf::{restriction_kind_with_profile, RestrictionKind, RestrictionProfile};
 
 /// Probability value in [0.0, 1.0]. Currently f64 for codebase consistency.
 /// Future: may migrate to f32 or bounded newtype for memory optimization.
@@ -416,186 +418,4 @@ async fn wire_single_batch_entry(
     .map_err(|e| format!("update_claim_belief: {e}"))?;
 
     Ok(())
-}
-
-/// Outcome of an edge-factor auto-wire pass.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum EdgeFactorOutcome {
-    /// Source claim has no stored interval — nothing to propagate.
-    SourceFactorless,
-    /// Relationship maps to `RestrictionKind::Neutral` — not an epistemic edge.
-    NonEpistemic,
-    /// Restriction produced a vacuous interval (no information transfer).
-    Vacuous,
-    /// BBA materialized and target belief recomputed.
-    Wired,
-}
-
-/// Auto-wire DS for an **epistemic edge** treated as a factor on the target claim.
-///
-/// Uses the existing `RestrictionKind` framework (per `epigraph-engine::sheaf`)
-/// to map the edge's relationship onto a transmission factor, applies it to
-/// the source's stored `EpistemicInterval`, materializes a BBA on the
-/// canonical binary frame, persists with `perspective_id = edge_id`, then
-/// re-combines all of the target's stored BBAs and updates its belief columns.
-///
-/// Returns `Ok(EdgeFactorOutcome::SourceFactorless)` when the source has no
-/// stored interval (NULL belief/plausibility on `claims` row); the caller can
-/// retry later once the source acquires a BBA.
-pub async fn auto_wire_ds_for_edge(
-    pool: &PgPool,
-    edge_id: Uuid,
-    edge_signer_agent_id: Uuid,
-    source_id: Uuid,
-    target_id: Uuid,
-    relationship: &str,
-) -> Result<EdgeFactorOutcome, String> {
-    let restriction =
-        restriction_kind_with_profile(relationship, &RestrictionProfile::scientific());
-    if matches!(restriction, RestrictionKind::Neutral) {
-        return Ok(EdgeFactorOutcome::NonEpistemic);
-    }
-
-    let source_row: Option<(Option<f64>, Option<f64>, Option<f64>)> =
-        sqlx::query_as("SELECT belief, plausibility, open_world_mass FROM claims WHERE id = $1")
-            .bind(source_id)
-            .fetch_optional(pool)
-            .await
-            .map_err(|e| format!("fetch source interval: {e}"))?;
-    let Some((Some(bel), Some(pl), ow_opt)) = source_row else {
-        return Ok(EdgeFactorOutcome::SourceFactorless);
-    };
-    let source_interval =
-        EpistemicInterval::new(bel, pl, ow_opt.unwrap_or((pl - bel).max(0.0) * 0.5));
-
-    let (restricted, source_strength) = match restriction {
-        RestrictionKind::Positive(f) => (restrict_epistemic_positive(&source_interval, f), f),
-        RestrictionKind::Negative(f) => (restrict_epistemic_negative(&source_interval, f), f),
-        RestrictionKind::FrameEvidence(f) => (
-            restrict_epistemic_frame_evidence(&source_interval, source_interval.betp(), f),
-            f,
-        ),
-        RestrictionKind::Neutral => unreachable!(),
-    };
-
-    if restricted.width() >= 0.999 && restricted.bel < 0.01 {
-        return Ok(EdgeFactorOutcome::Vacuous);
-    }
-
-    let frame_id = ensure_binary_frame(pool).await?;
-    let frame = binary_frame()?;
-    let bba = restricted
-        .to_mass_function(&frame)
-        .map_err(|e| format!("interval_to_bba: {e}"))?;
-    let masses_json = mass_to_json(&bba)?;
-
-    FrameRepository::assign_claim(pool, target_id, frame_id, Some(0))
-        .await
-        .map_err(|e| format!("assign_claim: {e}"))?;
-    PerspectiveRepository::ensure_edge_perspective(pool, edge_id, Some(edge_signer_agent_id))
-        .await
-        .map_err(|e| format!("ensure_edge_perspective: {e}"))?;
-
-    MassFunctionRepository::store_with_perspective(
-        pool,
-        target_id,
-        frame_id,
-        Some(edge_signer_agent_id),
-        Some(edge_id),
-        &masses_json,
-        None,
-        Some("edge_factor"),
-        Some(source_strength),
-        Some(relationship),
-    )
-    .await
-    .map_err(|e| format!("store BBA: {e}"))?;
-
-    recompute_combined_belief(pool, target_id, frame_id, &frame).await?;
-    Ok(EdgeFactorOutcome::Wired)
-}
-
-/// Re-fetch all BBAs on (claim, frame), discount by source_strength, combine,
-/// and write the resulting Bel/Pl/BetP/conflict/missing to the claims row.
-/// Mirrors the cascade in `auto_wire_ds_update` post-store.
-async fn recompute_combined_belief(
-    pool: &PgPool,
-    claim_id: Uuid,
-    frame_id: Uuid,
-    frame: &FrameOfDiscernment,
-) -> Result<(), String> {
-    let all_rows = MassFunctionRepository::get_for_claim_frame(pool, claim_id, frame_id)
-        .await
-        .map_err(|e| format!("get_for_claim_frame: {e}"))?;
-    if all_rows.is_empty() {
-        return Ok(());
-    }
-
-    let combined = if all_rows.len() == 1 {
-        let r = &all_rows[0];
-        let mf = parse_stored_bba(frame, &r.masses)?;
-        let reliability = r.source_strength.unwrap_or(1.0).clamp(0.0, 1.0);
-        combination::discount(&mf, reliability).map_err(|e| format!("discount: {e}"))?
-    } else {
-        let mut mass_fns = Vec::with_capacity(all_rows.len());
-        for row in &all_rows {
-            let mf = parse_stored_bba(frame, &row.masses)?;
-            let reliability = row.source_strength.unwrap_or(1.0).clamp(0.0, 1.0);
-            let d =
-                combination::discount(&mf, reliability).map_err(|e| format!("discount: {e}"))?;
-            mass_fns.push(d);
-        }
-        let (c, _) = combination::combine_multiple(&mass_fns, 0.9)
-            .map_err(|e| format!("combine_multiple: {e}"))?;
-        c
-    };
-
-    let (bel, pl, betp, conflict, missing) = compute_measures(&combined);
-    MassFunctionRepository::update_claim_belief(
-        pool,
-        claim_id,
-        bel,
-        pl,
-        conflict,
-        Some(betp),
-        missing,
-    )
-    .await
-    .map_err(|e| format!("update_claim_belief: {e}"))?;
-    Ok(())
-}
-
-/// Fire `auto_wire_ds_for_edge` from an edge-creation call site, gated on
-/// whether the edge was newly created and connects two claim nodes.
-///
-/// Best-effort: failures are logged at `warn` and swallowed. Returns `None`
-/// when the edge wasn't newly created, sources/targets aren't claims, or the
-/// auto-wire call failed. Returns `Some(outcome)` when the trigger fired.
-#[allow(clippy::too_many_arguments)]
-pub async fn auto_wire_edge_if_epistemic(
-    pool: &PgPool,
-    was_created: bool,
-    edge_id: Uuid,
-    source_id: Uuid,
-    source_type: &str,
-    target_id: Uuid,
-    target_type: &str,
-    relationship: &str,
-    agent_id: Uuid,
-) -> Option<EdgeFactorOutcome> {
-    if !was_created || source_type != "claim" || target_type != "claim" {
-        return None;
-    }
-    match auto_wire_ds_for_edge(pool, edge_id, agent_id, source_id, target_id, relationship).await {
-        Ok(outcome) => Some(outcome),
-        Err(e) => {
-            tracing::warn!(
-                edge = %edge_id,
-                target = %target_id,
-                relationship = %relationship,
-                "edge auto-wire failed: {e}",
-            );
-            None
-        }
-    }
 }
